Simplify ray collision calculating
